Chat: split finalizer tests to satisfy file-length gate
