+++ IntelligenceX.Chat/IntelligenceX.Chat.Tests/ChatServiceRoutingTrimTests.IntelligenceLoop.Integration.cs
@@ -357,66 +357,6 @@ public sealed partial class ChatServiceRoutingTrimTests {
     }
 
     [Fact]
-    public void FinalizePhaseHeartbeatFailure_DoesNothingWhenFailureIsNull() {
-        using var heartbeatCts = new CancellationTokenSource();
-        using var outerCts = new CancellationTokenSource();
-
-        ChatServiceSession.FinalizePhaseHeartbeatFailure(
-            heartbeatFailure: null,
-            phaseStatus: "phase_review",
-            requestId: "req-intelligence-loop",
-            threadId: "thread-intelligence-loop",
-            heartbeatCancellationToken: heartbeatCts.Token,
-            cancellationToken: outerCts.Token);
-    }
-
-    [Fact]
-    public void FinalizePhaseHeartbeatFailure_SuppressesExpectedIOException() {
-        using var heartbeatCts = new CancellationTokenSource();
-        using var outerCts = new CancellationTokenSource();
-
-        ChatServiceSession.FinalizePhaseHeartbeatFailure(
-            heartbeatFailure: new IOException("simulated-io"),
-            phaseStatus: "phase_review",
-            requestId: "req-intelligence-loop",
-            threadId: "thread-intelligence-loop",
-            heartbeatCancellationToken: heartbeatCts.Token,
-            cancellationToken: outerCts.Token);
-    }
-
-    [Fact]
-    public void FinalizePhaseHeartbeatFailure_RethrowsUnexpectedFailure() {
-        using var heartbeatCts = new CancellationTokenSource();
-        using var outerCts = new CancellationTokenSource();
-
-        var ex = Assert.Throws<InvalidOperationException>(() => ChatServiceSession.FinalizePhaseHeartbeatFailure(
-            heartbeatFailure: new InvalidOperationException("heartbeat-failed"),
-            phaseStatus: "phase_review",
-            requestId: "req-intelligence-loop",
-            threadId: "thread-intelligence-loop",
-            heartbeatCancellationToken: heartbeatCts.Token,
-            cancellationToken: outerCts.Token));
-
-        Assert.Equal("heartbeat-failed", ex.Message);
-    }
-
-    [Fact]
-    public void FinalizePhaseHeartbeatFailure_RethrowsCanceledFailureFromUnrelatedToken() {
-        using var heartbeatCts = new CancellationTokenSource();
-        using var outerCts = new CancellationTokenSource();
-        using var unrelatedCts = new CancellationTokenSource();
-        unrelatedCts.Cancel();
-
-        Assert.Throws<OperationCanceledException>(() => ChatServiceSession.FinalizePhaseHeartbeatFailure(
-            heartbeatFailure: new OperationCanceledException("unrelated", innerException: null, unrelatedCts.Token),
-            phaseStatus: "phase_review",
-            requestId: "req-intelligence-loop",
-            threadId: "thread-intelligence-loop",
-            heartbeatCancellationToken: heartbeatCts.Token,
-            cancellationToken: outerCts.Token));
-    }
-
-    [Fact]
     public async Task ToolRoundStatusLifecycle_EmitsRoundStatusesInDeterministicOrder() {
         var session = ChatServiceTestSessionFactory.CreateIsolatedSession();
         using var capture = new SynchronizedCaptureStream();
